chore: rename to contract
